Fix to #19499 - Query: null semantics not applied on subquery.Contains(null)
We used to translate this into `NULL IN subquery` pattern, but this doesn't work because it doesn't match if the subquery also contains null.
Fix is to convert this into subquery.Any(e => e == NULL), which translates to EXISTS with predicate and we can correctly apply null semantics.
Also it allows us to translate contains on entities with composite keys.

Also made several small fixes:
- marked EXISTS and IN expressions as never nullable for purpose of null semantics,
- optimized EXISITS (subquery) with predicate that resolves to false, directly into false, since empty subquery never exisits,
- improves expression printer output for IN expression in the subquery scenario.